Generic delimit strainer
Refactored delimit strainer into two subclasses to avoid having to
figure out how to pass in arguments. Better if users specify and name,
and the rest is taken care of (rather than specify the name of the
strainer, along with parameters to pass it).

Now class specifically for multiple space replacement, and new one for
replacing spaces with underscores (for n-gram parsing).