Ability to keep sentence endings.
Part of the reason the query models do not behave the way they did in
the NAACL paper is because there are too many overlapping regions per
document. If there were fewer, it may create "silence" that was seen
in the audio output, and which was key for the algorithms to
stop. Keeping the punctuation may allow output that is more inline
with the audio.