Corrected strainer recursion. Character deletion replaces.
Strain only if the strainer has been set. Returning a xerox'd document
is the wrong thing to do: send the xerox to the strainer, which will
eventally return it. (Manipulation is the text-based middleman.)

When remove characters don't set the table entry to None. Instead add
a space as later strainers know how to deal with those sensibly.