Off-by-one error when interpreting Aren's output.
For a Term, the length of its n-gram and where the Term ends in the
document are two different things; the former being one larger than
the difference of the latter from the start. To take this into
account, the arbitrary n-gram that is created from Aren's output must
have an additional character based on its reported length. This allows
the concept of the Term.span to align with the concept of Aren's term
duration.